Calculus Study Aid
In college, I majored in Mathematics. As a result, one would think I'd
know Calculus backwards and forwards, but it's been several years and
my grasp is embarrasingly poor.

There are a lot of great resources to help somebody learn calculus:
EdX, OCW, Khan Academy, etc. The problem is, none of them are very
good at solving my unique set of problems:

* My knowledge is patchy, there are some things I remember well and
  others that I'll recall once I've looked over them a bit.
* I don't want to find myself in this position again in ten years.

An effective approach to efficiently retaining this information is
spaced repetition. SuperMemo, Anki and Mnemosyne are all great ways to
do this with specific facts, but personally I haven't found them
particularly effective in helping me realize to apply them in
problems.

What I really want is a collection of problems, for which any one is
some combination of concepts. Spaced repetition is then applied to the
concepts, so I'm not spinning my wheels on a bunch of easy problems
for which I clearly understand what's going on. A nice to have here
would be a link to an OCW lecture clip explaining the concept.

I'd also like to eventually extend this to other classes. Ultimately,
this should be a tool to retain subject matter in the most efficient
manner possible. I could also see it serving as a study aid for
students currently in a class.