Routing between even function and finished
Not the right algorithm when I get there, but have the study url
redirect to the next thing to study. I thought about just rendering
everything at /study, but concluded that if I ever wanted to pass a
link to a problem having that information in the url would be useful.

I also learned the difference between a 301 and 303 redirect. I
started with a 301 and then spent a bunch of time scratching my head
as to why the redirect wasn't changing. Since it's considered to be
moved permanently, browsers take that as license to cache the
subsequent location and send you directly there.