Generate Even Function
Wrote code to generate even functions based on a single exponent. One
design choice I settled on is always making what you see linkable,
which leads to multiple redirects. Not too concerned about this as the
only user, but I may need something to short-circuit to one redirect
later on.

My idea here is, if I have trouble with a problem, I want to be able
to send it to somebody else and ask their opinion. My main goal right
now is studying in an efficient way, but I could see different people
using this in different ways later on.