Fix Routing
I guess I introduced some routing bugs trying to handle existing and
nonexistent path params in the same logic. Split the concepts out for
now, although I might consolidate at some point in the future.

I also had the line breaks wrong in the README for what emacs wants,
so I redid that. In addition, the `restart` function was only
reloading the server, and not all of its dependencies.