SM-2 Correct Data Structures
Still working on actually implementing SM-2, but the /study route now
sets data structures up correctly. Before we were just toggling
continue back and forth, but what study will need to do in the long
run is determine which concepts need to be reviewed. This incorporated
dates, which was more complicated than I had hoped, so I ended up
stopping after getting data structure handling finished.

The general idea here is that there is a set of concepts to
review, and this accumulates more as we reach new dates. There are
three main cases to worry about here:

* Last review date doesn't exist: add all concepts for review
* Last review date is in the past: add any concepts for days between
  then and now
* Last review date is today: do not add any concepts

Once these cases have been addressed, a concept is randomly selected
for the redirect. Concepts for past days are stored as a set on that
day. That's not being added anywhere yet, but the logic to pull that
information from redis is in place.

I had to deal with some low level stuff in redis and date as well. For
redis, I was unsatisfied with the naming scheme and chose to switch
to using `get` and `set` and just excluding those in that namespace. I
also did some type examining in `set` to make sets be stored as such
in redis. With dates, I mostly just did a thin wrapper over java.time,
although I did set up an infinite sequence that I could see being used
more in the future.