Implement SM-2 For Even Functions
To begin with, I just wanted to get SM-2 into the codebase. As I
mentioned in the readme, I plan on heavily modifying the algorithm,
but I wanted a starting point close to what's described in
https://www.supermemo.com/english/ol/sm2.htm.

A couple things I learned about clojure and pedestal along the way:

* Pedestal doesn't automatically include the form body in a request, I
  had to use the body-params intercepter to make that happen.
* Clojure doesn't consider sets a seq

On top of these changes, I introduced the beginnings of type retention
coming in and out of redis. I don't want to be constantly doing
`read-string` or the like when I call get, so I'm storing what keys
are numbers so the conversion happens automatically.